Pass string to "parent" field when creating or moving a page.

+++ test/Permissions/Editor/Create/AllowedTest.php
@@ -47,7 +47,9 @@ class EditorCreateAllowedOnlyOneTest extends GraphqlTestCase {
   public static function setUpBeforeClass()
   {
     $searchPage = Utils::pages()->get("template=search");
-    $searchPage->delete();
+    if ($searchPage->id) {
+      $searchPage->delete();
+    }
     parent::setUpBeforeClass();
   }
 
@@ -63,7 +65,7 @@ class EditorCreateAllowedOnlyOneTest extends GraphqlTestCase {
 
     $variables = [
       'page' => [
-        'parent' => 1,
+        'parent' => '1',
         'name' => 'search',
         'title' => 'Search'
       ]
